renamed various image files and updated path
