+++ src/components/hero.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import photo from '../assets/images/photos/IMG_0073.jpg';
+import photo from '../assets/images/photos/photo_73.jpg';
 
 function Hero() {
   return (
